requirements.txt
aniso8601==8.0.0
apispec==3.1.0
atomicwrites==1.3.0
attrs==19.3.0
Click==7.0
Flask==1.1.1
flask-apispec==0.8.3
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.7
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
greenlet==0.4.15
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.2.2
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.2.3
pytest-env==0.6.2
pytest-ordering==0.6
python-dateutil==2.8.1
python-engineio==3.11.1
python-socketio==4.4.0
pytz==2019.3
redis==3.3.11
six==1.13.0
SQLAlchemy==1.3.11
wcwidth==0.1.7
webargs==5.5.2
Werkzeug==0.16.0
zipp==0.6.0
